+++ tests/testthat/test_statistics.R
@@ -1,3 +1,41 @@
+
+test_that("poor_man_stack() works (test could be improved", {
+  y <- c("a", "b", "c")
+  z <- c("aa", "bb", "cc")
+  X <- data.frame(x = 1:3, y = y, z = z)
+  out <- poor_man_stack(X, to_stack = c("y", "z"))
+  xpected <- data.frame(
+    x = rep(1:3, times = 2L), 
+    varying_ = factor(rep(c("y", "z"), each = 3L)),
+    value_ = c(y, z)
+  )
+  expect_equal(out, xpected)
+  
+  expect_error(poor_man_stack(cbind(a = 1:3, b = 2:4), to_stack = "b"))
+})
+
+test_that("mat2df() works (test could be improved)", {
+  mat <- cbind(y = 1:2, z = c(0.5, 0.5))
+  rownames(mat) <- letters[seq_len(nrow(mat))]
+  out <- mat2df(mat)
+  rownames(out) <- NULL
+  xpected <- data.frame(
+    id_ = "Overall", 
+    variable_ = factor(c("a", "b", "a", "b")),
+    varying_ = factor(c("y", "y", "z", "z")),
+    value_ = c(1, 2, 0.5, 0.5),
+    stringsAsFactors = FALSE
+  )
+  expect_equal(out, xpected)
+  
+  mat_no_names <- mat
+  colnames(mat_no_names) <- NULL
+  expect_equal(unique(mat2df(mat_no_names)$varying_), factor(c("y1", "y2")))
+  
+  expect_error(mat2df(head(iris)))
+  expect_error(mat2df(1:4))
+})
+
 test_that("postprocess() works for matrix input", {
   num <- cbind(a = 1:3, b = c(1, 1, 1))
   denom <- cbind(a = 1:3, b = 1:3)